refactor: move images to images file
