+++ src/pages/Admin.jsx
@@ -1,0 +1,8 @@
+function Admin() {
+  return (
+    <div>
+      <h1>Admin</h1>
+    </div>
+  );
+}
+export default Admin;
